Refactor Popup&banner controller and Popup&banner models validations #227
