+++ backend/src/controllers/banner.controller.js
@@ -1,18 +1,9 @@
 const bannerService = require("../service/banner.service.js");
-const { internalServerError } = require("../utils/errors");
-const { isValidHexColor, checkColorFields } = require("../utils/guideHelpers");
-const db = require("../models"); 
+const { internalServerError } = require("../utils/errors.js");
+const { isValidHexColor, checkColorFields } = require("../utils/guideHelpers.js");
+const { validatePosition, validateCloseButtonAction } = require("../utils/banner.helper.js");
+const db = require("../models/index.js"); 
 const Banner = db.Banner;
-
-const validatePosition = (value) => {
-  const validPositions = ["top", "bottom"];
-  return validPositions.includes(value);
-};
-
-const validateCloseButtonAction = (value) => {
-  const validActions = ["no action", "open url", "open url in a new tab"];
-  return validActions.includes(value);
-};
 
 class BannerController {
   async addBanner(req, res) {
@@ -85,7 +76,7 @@ class BannerController {
   async editBanner(req, res) {
     try {
       const { id } = req.params;
-
+  
       if (!req.body.position || !req.body.closeButtonAction) {
         return res
           .status(400)
@@ -93,38 +84,22 @@ class BannerController {
             errors: [{ msg: "position and closeButtonAction are required" }],
           });
       }
-
-      const positionColumn = Banner.tableAttributes.position;
-
-      if (!positionColumn.validate.isIn[0].includes(req.body.position)) {
+  
+      if (!validatePosition(req.body.position)) {
         return res
           .status(400)
           .json({ errors: [{ msg: "Invalid value for position" }] });
       }
-
-      const closeButtonActionColumn = Banner.tableAttributes.closeButtonAction;
-
-      if (
-        !closeButtonActionColumn.validate.isIn[0].includes(
-          req.body.closeButtonAction,
-        )
-      ) {
+  
+      if (!validateCloseButtonAction(req.body.closeButtonAction)) {
         return res
           .status(400)
           .json({ errors: [{ msg: "Invalid value for closeButtonAction" }] });
       }
-
-      const colorFields = ["fontColor", "backgroundColor"];
-      for (const field of colorFields) {
-        if (req.body[field] && !isValidHexColor(req.body[field])) {
-          return res
-            .status(400)
-            .json({
-              errors: [{ msg: `${field} must be a valid hex color code` }],
-            });
-        }
-      }
-
+  
+      const colorFields = { fontColor: req.body.fontColor, backgroundColor: req.body.backgroundColor };
+      checkColorFields(colorFields, res);
+  
       const updatedBanner = await bannerService.updateBanner(id, req.body);
       res.status(200).json(updatedBanner);
     } catch (err) {
@@ -148,6 +123,7 @@ class BannerController {
       res.status(statusCode).json(payload);
     }
   }
+  
   async getBanners(req, res) {
     try {
       const userId = req.user.id;
